Add the ability to specify a router_url
Adds the ability to specify a router_url, therefore skipping the UPNP
discovery step entirely. This makes it possible to use this daemon in a
network, where the router is located in a different subnet.
For example this daemon might be deployed as container and therefore use
a different network than the router.